cleanup shell prompt
Uses the fill module instead of right prompt in order to fix my
annoyance with the right prompt always appearing on the same line as the
input. The right prompt will always show up when copying a command its
output. This is a workaround. Instead of using a right prompt, the fill
module pushes stuff after it all the way to the right.

Signed-off-by: Daniel Schaaff <daniel@danielschaaff.com>